# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Supply Chain – Micro-Warehouse Network for Resilient Logistics (India)
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  admin: "admin",
  manager: "manager",
  retail: "retail",
  customer: "customer",
} as const;

export type LocationType = (typeof LocationType)[keyof typeof LocationType];

export const LocationType = {
  warehouse: "warehouse",
  micro: "micro",
} as const;

export type StockStatus = (typeof StockStatus)[keyof typeof StockStatus];

export const StockStatus = {
  healthy: "healthy",
  low: "low",
  critical: "critical",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  dispatched: "dispatched",
  in_transit: "in_transit",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: Role;
  city: string;
  createdAt: string;
}

export interface LoginBody {
  email: string;
}

export interface MeResponse {
  user: User | null;
}

export interface Product {
  id: number;
  name: string;
  sku: string;
  category: string;
  unitPrice: number;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  sku: string;
  category: string;
  unitPrice: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface UpdateProductBody {
  name?: string;
  sku?: string;
  category?: string;
  unitPrice?: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Location {
  id: number;
  name: string;
  type: LocationType;
  city: string;
  state: string;
  address: string;
  lat: number;
  lng: number;
  capacity: number;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  createdAt: string;
}

export type LocationWithStats = Location & {
  totalUnits: number;
  distinctSkus: number;
  stockStatus: StockStatus;
  utilizationPct: number;
};

export interface InventoryRow {
  id: number;
  productId: number;
  locationId: number;
  quantity: number;
  threshold: number;
  productName: string;
  productSku: string;
  productCategory: string;
  locationName: string;
  locationType: LocationType;
  locationCity: string;
  locationState: string;
  stockStatus: StockStatus;
  updatedAt: string;
}

export type LocationDetail = LocationWithStats & {
  inventory: InventoryRow[];
};

export interface CreateLocationBody {
  name: string;
  type: LocationType;
  city: string;
  state: string;
  address: string;
  lat: number;
  lng: number;
  capacity: number;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  contactPhone?: string | null;
}

export interface UpdateLocationBody {
  name?: string;
  type?: LocationType;
  city?: string;
  state?: string;
  address?: string;
  lat?: number;
  lng?: number;
  capacity?: number;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  contactPhone?: string | null;
}

export interface CreateInventoryBody {
  productId: number;
  locationId: number;
  quantity: number;
  threshold: number;
}

export interface UpdateInventoryBody {
  quantity?: number;
  threshold?: number;
}

export type ShortageRow = InventoryRow & {
  shortfall: number;
  /** @nullable */
  nearestAlternativeName?: string | null;
  /** @nullable */
  nearestAlternativeDistanceKm?: number | null;
};

export interface SuggestFulfillmentBody {
  productId: number;
  quantity: number;
  deliveryLat: number;
  deliveryLng: number;
}

export interface FulfillmentSuggestion {
  locationId: number;
  locationName: string;
  locationType: LocationType;
  city: string;
  state: string;
  lat: number;
  lng: number;
  availableQuantity: number;
  distanceKm: number;
  etaMinutes: number;
  score: number;
}

export interface Order {
  id: number;
  customerId: number;
  customerName: string;
  productId: number;
  productName: string;
  productSku: string;
  quantity: number;
  /** @nullable */
  sourceLocationId?: number | null;
  /** @nullable */
  sourceLocationName?: string | null;
  sourceLocationType?: LocationType | null;
  deliveryAddress: string;
  deliveryCity: string;
  deliveryLat: number;
  deliveryLng: number;
  distanceKm: number;
  etaMinutes: number;
  status: OrderStatus;
  totalPrice: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderBody {
  customerId: number;
  productId: number;
  quantity: number;
  deliveryAddress: string;
  deliveryCity: string;
  deliveryLat: number;
  deliveryLng: number;
}

export interface UpdateOrderStatusBody {
  status: OrderStatus;
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  info: "info",
  warning: "warning",
  critical: "critical",
} as const;

export interface Alert {
  id: number;
  type: string;
  severity: AlertSeverity;
  message: string;
  /** @nullable */
  locationId?: number | null;
  /** @nullable */
  locationName?: string | null;
  /** @nullable */
  productId?: number | null;
  /** @nullable */
  productName?: string | null;
  resolved: boolean;
  createdAt: string;
}

export type AnalyticsSummaryOrdersByStatusItem = {
  status: OrderStatus;
  count: number;
};

export interface AnalyticsSummary {
  totalLocations: number;
  totalWarehouses: number;
  totalMicroWarehouses: number;
  totalProducts: number;
  totalUnits: number;
  totalOrders: number;
  ordersToday: number;
  activeAlerts: number;
  criticalShortages: number;
  avgDeliveryMinutes: number;
  avgDistanceKm: number;
  microWarehouseSharePct: number;
  revenue: number;
  ordersByStatus: AnalyticsSummaryOrdersByStatusItem[];
}

export interface TrendPoint {
  date: string;
  totalUnits: number;
  ordersCount: number;
  shortagesCount: number;
}

export interface RegionDemand {
  state: string;
  ordersCount: number;
  unitsSold: number;
  revenue: number;
}

export interface HeatmapCell {
  state: string;
  city: string;
  totalUnits: number;
  locationsCount: number;
  shortagesCount: number;
  stockStatus: StockStatus;
}

export type DeliveryEfficiencyByDayItem = {
  date: string;
  avgEtaMinutes: number;
  ordersCount: number;
};

export interface DeliveryEfficiency {
  avgDistanceKm: number;
  avgEtaMinutes: number;
  ordersFromMicro: number;
  ordersFromWarehouse: number;
  onTimePct: number;
  byDay: DeliveryEfficiencyByDayItem[];
}

export type ProductForecastForecastItem = {
  date: string;
  expectedDemand: number;
};

export interface ProductForecast {
  productId: number;
  productName: string;
  category: string;
  currentStock: number;
  avgDailyDemand: number;
  forecast: ProductForecastForecastItem[];
  /** @nullable */
  daysUntilStockout?: number | null;
}

export type AiChatMessageRole =
  (typeof AiChatMessageRole)[keyof typeof AiChatMessageRole];

export const AiChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface AiChatMessage {
  role: AiChatMessageRole;
  content: string;
}

export interface AiChatBody {
  message: string;
  history?: AiChatMessage[];
}

export interface AiChatResponse {
  reply: string;
  sources: string[];
}

export type ListLocationsParams = {
  type?: ListLocationsType;
};

export type ListLocationsType =
  (typeof ListLocationsType)[keyof typeof ListLocationsType];

export const ListLocationsType = {
  warehouse: "warehouse",
  micro: "micro",
} as const;

export type ListInventoryParams = {
  locationId?: number;
  productId?: number;
};

export type ListOrdersParams = {
  customerId?: number;
};

export type GetInventoryTrendsParams = {
  days?: number;
};