# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Supply Chain – Micro-Warehouse Network for Resilient Logistics (India)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login by email (demo)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginBody" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List locations with inventory snapshot
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [warehouse, micro]
      responses:
        "200":
          description: Locations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LocationWithStats" }
    post:
      operationId: createLocation
      tags: [locations]
      summary: Create location
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateLocationBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Location" }
  /locations/{id}:
    get:
      operationId: getLocation
      tags: [locations]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Location with inventory
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LocationDetail" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
    patch:
      operationId: updateLocation
      tags: [locations]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateLocationBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Location" }
    delete:
      operationId: deleteLocation
      tags: [locations]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }

  # -----------------------------------------------------------------------
  # Inventory
  # -----------------------------------------------------------------------
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: All inventory rows (joined product + location)
      parameters:
        - name: locationId
          in: query
          schema: { type: integer }
        - name: productId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Inventory rows
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InventoryRow" }
    post:
      operationId: createInventory
      tags: [inventory]
      summary: Place product at a location
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInventoryBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryRow" }
  /inventory/{id}:
    patch:
      operationId: updateInventory
      tags: [inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateInventoryBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryRow" }
    delete:
      operationId: deleteInventory
      tags: [inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
...
  /alerts/{id}/resolve:
    patch:
      operationId: resolveAlert
      tags: [alerts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Alert" }

  # -----------------------------------------------------------------------
  # Analytics (wow endpoints)
  # -----------------------------------------------------------------------
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error: { type: string }
      required: [error]
    Role:
      type: string
      enum: [admin, manager, retail, customer]
    LocationType:
      type: string
      enum: [warehouse, micro]
    StockStatus:
      type: string
      enum: [healthy, low, critical]
...
    User:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        city: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, email, name, role, city, createdAt]
    LoginBody:
      type: object
      properties:
        email: { type: string }
      required: [email]
...
    Location:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        type: { $ref: "#/components/schemas/LocationType" }
        city: { type: string }
        state: { type: string }
        address: { type: string }
        lat: { type: number }
        lng: { type: number }
        capacity: { type: integer }
        contactName:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
      required: [id, name, type, city, state, address, lat, lng, capacity, createdAt]
    LocationWithStats:
      allOf:
        - $ref: "#/components/schemas/Location"
        - type: object
          properties:
            totalUnits: { type: integer }
            distinctSkus: { type: integer }
            stockStatus: { $ref: "#/components/schemas/StockStatus" }
            utilizationPct: { type: number }
          required: [totalUnits, distinctSkus, stockStatus, utilizationPct]
    LocationDetail:
      allOf:
        - $ref: "#/components/schemas/LocationWithStats"
        - type: object
          properties:
            inventory:
              type: array
              items: { $ref: "#/components/schemas/InventoryRow" }
          required: [inventory]
    CreateLocationBody:
      type: object
      properties:
        name: { type: string }
        type: { $ref: "#/components/schemas/LocationType" }
        city: { type: string }
        state: { type: string }
        address: { type: string }
        lat: { type: number }
        lng: { type: number }
        capacity: { type: integer }
        contactName:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
      required: [name, type, city, state, address, lat, lng, capacity]
    UpdateLocationBody:
      type: object
      properties:
        name: { type: string }
        type: { $ref: "#/components/schemas/LocationType" }
        city: { type: string }
        state: { type: string }
        address: { type: string }
        lat: { type: number }
        lng: { type: number }
        capacity: { type: integer }
        contactName:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
    InventoryRow:
      type: object
      properties:
        id: { type: integer }
        productId: { type: integer }
        locationId: { type: integer }
        quantity: { type: integer }
        threshold: { type: integer }
        productName: { type: string }
        productSku: { type: string }
        productCategory: { type: string }
        locationName: { type: string }
        locationType: { $ref: "#/components/schemas/LocationType" }
        locationCity: { type: string }
        locationState: { type: string }
        stockStatus: { $ref: "#/components/schemas/StockStatus" }
        updatedAt: { type: string, format: date-time }
      required:
        [id, productId, locationId, quantity, threshold, productName, productSku,
         productCategory, locationName, locationType, locationCity, locationState,
         stockStatus, updatedAt]
    CreateInventoryBody:
      type: object
      properties:
        productId: { type: integer }
        locationId: { type: integer }
        quantity: { type: integer }
        threshold: { type: integer }
      required: [productId, locationId, quantity, threshold]
    UpdateInventoryBody:
      type: object
      properties:
        quantity: { type: integer }
        threshold: { type: integer }
...
    Alert:
      type: object
      properties:
        id: { type: integer }
        type: { type: string }
        severity:
          type: string
          enum: [info, warning, critical]
        message: { type: string }
        locationId:
          type: ["integer", "null"]
        locationName:
          type: ["string", "null"]
        productId:
          type: ["integer", "null"]
        productName:
          type: ["string", "null"]
        resolved: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, type, severity, message, resolved, createdAt]
...