# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Supply Chain – Micro-Warehouse Network for Resilient Logistics (India)
 * OpenAPI spec version: 0.1.0
 */

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  admin: "admin",
  manager: "manager",
  retail: "retail",
  customer: "customer",
} as const;